Internal heplots functions

Internal functions for package heplots.

lambda.crit(alpha, p, dfh, dfe, test.statistic=c("Roy", "HLT", "Hotelling-Lawley"))
Roy.crit(alpha, p, dfh, dfe)
HLT.crit(alpha, p, dfh, dfe)
he.rep(x, n)
termInfo(x)
last(x)

Arguments

alpha
significance level for critical values of multivariate statistics

p
Number of variables

dfh
degrees of freedom for hypothesis

dfe
degrees of freedom for error

test.statistic
Test statistic used for the multivariate test

x
An argument to heplot or heplot3d that is to be repeated for Error and all hypothesis terms

n
Number of hypothesis terms

Details

These functions are not intended to be called by the user.