Effect of Delay in Oral Practice in Second Language Learning

Postovsky (1970) investigatged the effect of delay in oral practice at the beginning of second language learning. A control condition began oral practice with no delay, while an experimental group had a four-week delay before starting oral practice. The data consists of scores on language skills at the end of six weeks of study.

Students in this study were matched on age, education, former language training, intelligence and language aptitude.

Usage

data("oral")

Format

A data frame with 56 observations on the following 5 variables.

group

Group, a factor with levels Control Exptl

listen

Listening test, a numeric vector

speak

Speaking test, a numeric vector

read

Reading test, a numeric vector

write

Writing test, a numeric vector

Source

Timm, N. H. (1975). Multivariate Analysis with Applications in Education and Psychology. Wadsworth (Brooks/Cole), Exercise 3.12, p. 279.

References

Postovsky, V. A. (1970). Effects of delay in oral practice at the start of second language training. Unpublished doctoral dissertation, University of California, Berleley.

Examples

data(oral)
## maybe str(oral) ; plot(oral) ...