heplots-internal(heplots)	R Documentation

Internal heplots functions

Description

Internal functions for package heplots.

Usage


lambda.crit(alpha, p, dfh, dfe, test.statistic=c("Roy", "HLT", "Hotelling-Lawley"))
Roy.crit(alpha, p, dfh, dfe)
HLT.crit(alpha, p, dfh, dfe)
he.rep(x, n)
termInfo(x)
termMeans(mod, term, label.factors=TRUE)

Arguments

alpha

p 	 Number of variables

dfh 	 degrees of freedom for hypothesis

dfe 	 degrees of freedom for error

test.statistic 	 Test statistic used for the multivariate test

x 	 An argument to heplot or heplot3d that is to be repeated for Error and all hypothesis terms

n 	 Number of hypothesis terms

Details

These functions are not intended to be called by the user.

Author(s)

Michael Friendly friendly@yorku.ca

[Package heplots version 0.9-1 Index]